refactor(server): split MCP proxy into phase modules
proxy.ts (1,922 lines) keeps the McpProxy class and public surface; shared
helpers, upstream client (session cache + credential resolution + JSON-RPC
egress), auto-auth flows, REST handlers and the streamable-HTTP forward path
move to proxy-shared/proxy-upstream/proxy-auth-flows/proxy-rest-routes/
proxy-forward. Token/credential/SSRF/approval logic byte-identical; tokenData
anys replaced with the actual WorkerTokenData type.
